Moved collision logic to own helper per bar renderer
